"""fix of the review below : Decided to use the "int" type for the references, this means when there are more than 2.1 billion transactions his application will stop working Solution : Replaced int with long
"""

+++ src/main/java/nl/rabobank/processor/api/model/response/FailedRecordResponse.java
@@ -3,6 +3,6 @@ package nl.rabobank.processor.api.model.response;
 import lombok.Data;
 @Data
 public class FailedRecordResponse {
-    private int reference;
+    private long reference;
     private String description;
 }
